run against new disabled test and move stories under form
