+++ src/js/components/Form/stories/FieldStatesCustom.js
@@ -2,8 +2,65 @@ import React, { useRef, useEffect } from 'react';
 import { storiesOf } from '@storybook/react';
 
 import { grommet, Box, Form, FormField, TextInput, Grommet } from 'grommet';
+import { deepMerge } from '../../../utils';
 
-const FormFieldStates = () => {
+const customTheme = deepMerge(grommet, {
+  formField: {
+    border: {
+      error: {
+        color: 'border',
+      },
+      color: 'border',
+      side: 'all',
+    },
+    disabled: {
+      background: {
+        color: undefined,
+      },
+      border: {
+        color: 'status-disabled',
+      },
+      label: {
+        color: 'status-disabled',
+      },
+    },
+    error: {
+      background: {
+        color: { light: '#FF404033', dark: '#FF40404D' },
+      },
+      size: 'xsmall',
+      color: 'text-weak',
+      margin: {
+        start: 'none',
+      },
+    },
+    help: {
+      size: 'xsmall',
+      color: 'text-weak',
+      margin: {
+        start: 'none',
+        bottom: 'xsmall',
+      },
+    },
+    info: {
+      size: 'xsmall',
+      color: 'text-weak',
+      margin: {
+        start: 'none',
+      },
+    },
+    label: {
+      size: 'xsmall',
+      color: 'text-weak',
+      margin: {
+        horizontal: 'none',
+      },
+    },
+    round: '4px',
+  },
+});
+
+const CustomFormFieldStates = () => {
   const inputRef = useRef();
 
   useEffect(() => {
@@ -11,7 +68,7 @@ const FormFieldStates = () => {
   }, []);
 
   return (
-    <Grommet theme={grommet}>
+    <Grommet theme={customTheme}>
       <Box align="center" pad="large">
         <Form>
           <Box border gap="medium" pad="large" width="medium">
@@ -76,4 +133,6 @@ const FormFieldStates = () => {
   );
 };
 
-storiesOf('FormField', module).add('Field States', () => <FormFieldStates />);
+storiesOf('Form', module).add('Field States Custom', () => (
+  <CustomFormFieldStates />
+));
