+++ src/js/components/FormField/__tests__/FormField-test.js
@@ -221,14 +221,6 @@ describe('FormField', () => {
     expect(tree).toMatchSnapshot();
   });
 
-  // This test should be fully un-commented for PR #3973
-  // Before un-commenting this test fully, merge this test with branch
-  // halocline:formfield-disabled-styling an run `yarn test` to
-  // make sure the snapshots passes as is, and
-  // that should confirm the backward compatibility aspect.
-  // If that passes as expected, continue with
-  // un-commenting the theme properties of disabled, and verify that the only 
-  // snapshot change is for the text color (changing from red to teal)
   test('disabled with custom label', () => {
     const component = renderer.create(
       <Grommet
@@ -240,11 +232,11 @@ describe('FormField', () => {
               margin: 'xsmall',
               weight: 600,
             },
-            // disabled: {
-            //   label: {
-            //     color: 'teal',
-            //   },
-            // },
+            disabled: {
+              label: {
+                color: 'teal',
+              },
+            },
           },
         }}
       >
